requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary  # PostgreSQL driver
redis  # Redis caching (optional)
pytest  # Unit testing
httpx  # Required for FastAPI TestClient
python-dotenv  # Environment variable management
jinja2
fastapi[all]